Update authoring_test to not run with python3 anymore
The authoring tests now include cnx-archive which is not python3
compatible.  So not going to run the tests with python3 for now.

Also add a sed command for changing storage.conn.dsn to the original
connection string in the settings file.  The reason is that my
postgresql setup requires a password even for local access and
storage.conn.dsn does not have the password.